#25 Added a bunch of stuff to AssumptionPane + reformat project code

+++ src/main/java/sh/shinterface/datacontainer/Deck.java
@@ -2,10 +2,10 @@ package sh.shinterface.datacontainer;
 
 public record Deck(int allLib, int allFasc, int restLib, int restFasc) {
 
-    public float predictChance(int numberOfLibs) {
-        int teller = 0;
-        int restTotal = restLib + restFasc;
-        int noemer = restTotal * (restTotal - 1) * (restTotal - 2);
+    public double predictChance(int numberOfLibs) {
+        double teller = 0;
+        double restTotal = restLib + restFasc;
+        double noemer = restTotal * (restTotal - 1) * (restTotal - 2);
         if (numberOfLibs == 0) {
             teller = restFasc * (restFasc - 1) * (restFasc - 2);
         } else if (numberOfLibs == 1) {
@@ -15,6 +15,6 @@ public record Deck(int allLib, int allFasc, int restLib, int restFasc) {
         } else if (numberOfLibs == 3) {
             teller = restLib * (restLib - 1) * (restLib - 2);
         }
-        return (float) teller / (float) noemer;
+        return teller / noemer;
     }
 }
